Day 5 and some other fixes

+++ day2/main.cpp
@@ -4,7 +4,7 @@
 inline static bool isSafe(const std::vector<int> nums) {
   char dir = nums[1] > nums[0] ? 'a' : 'd';
 
-  for (int i = 1; i < nums.size(); i++) {
+  for (unsigned long i = 1; i < nums.size(); i++) {
     int curr = nums[i];
     int prev = nums[i - 1];
     int diff = abs(curr - prev);
@@ -23,10 +23,10 @@ inline static bool isSafeWithDampener(const std::vector<int>& original) {
   if (isSafe(original))
     return true;
 
-  for (int skip = 0; skip < original.size(); skip++) {
+  for (unsigned long skip = 0; skip < original.size(); skip++) {
     std::vector<int> modified;
 
-    for (int i = 0; i < original.size(); i++) {
+    for (unsigned long i = 0; i < original.size(); i++) {
       if (i != skip) {
         modified.push_back(original[i]);
       }
